func/requirements.txt
azure-functions
azure-identity
azure-mgmt-costmanagement
azure-mgmt-resource
msgraph-sdk